main関数を消した
