+++ docs/3-web-servers/09-cookie/_samples/profile/prisma/schema.prisma
