+++ docs/3-web-servers/09-cookie/_samples/profile/resister.ejs
